Huge rewrite to work better with cache and index.
The code was restructured for a better, in my option, code organisation.
Now it is possible to fetch the `index.json` file but it is still not
working with `--list` option.
Before checking for a command page it look in the index for the paltform
for that command, this way we avoid try fetching a page from a platform
  where it doesn't exists.